requirements.txt
openai == 1.33.0
langchain == 0.2.3
streamlit == 1.35.0
python-dotenv == 1.0.1
langchain-community == 0.2.0